+++ lib/utils/cert/selfsigned.go
@@ -16,7 +16,7 @@
  * /
  */
 
-package tls
+package cert
 
 import (
 	"crypto/rand"
